+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 Features:
   - Test Run List: Add posibility to filter by Inactive Milestone -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/73)
+  - Mark import af failed when import is failed  -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/79)
 
 Bugfixes:
   - Test Runs: Fix Filter for No Resolution Column -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/22)
+  - Test Run View: Chart is not updated when user updating results in a bulk -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/78)
   
 ## 0.3.7 (2020-03-02)
 
@@ -18,6 +20,9 @@ Features:
   - Milestone: Add Due Date to Milestone. -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/61)
   - Milestone: Add possibility to Close Milestone. -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/62)
   - Test Runs List: Add possibility to Add milestone from Test Runs List. -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/59)
+
+  Bugfixes:
+  - Fixed for downloading audits results from Audits Dashboard -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/51)
 
 ## 0.3.6 (2020-02-24)